[Cleanup] rename chord_tester to chord_reader
